+++ src/util.js
@@ -4,12 +4,14 @@ module.exports = {
   digestDefinitionFnOutput: function digestDefinitionFnOutput(output) {
     var vtree$;
     var onMount;
+    var dispose;
     var customEvents = {};
     if (output && output.hasOwnProperty('vtree$') &&
       typeof output.vtree$.subscribe === 'function')
     {
       vtree$ = output.vtree$;
       onMount = output.onMount;
+      dispose = output.dispose;
       customEvents = output;
     } else if (output && typeof output.subscribe === 'function') {
       vtree$ = output;
@@ -22,6 +24,7 @@ module.exports = {
     return {
       vtree$: vtree$,
       onMount: onMount,
+      dispose: dispose,
       customEvents: customEvents
     };
   }